# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Shift-Left Review Orchestrator
  description: 'Local PR review API on the operator host. Findings and policy decisions
    require human review and are not compliance verdicts. Advisory model findings
    are omitted from POST /api/v1/review by default; retrieve them from GET /api/v1/findings/{owner}/{repo}/{pr_number}
    or GET /api/v1/pr/{owner}/{repo}/{pr_number}/context. There is no Forgejo webhook
    listener on this process — Actions jobs call POST /api/v1/review with a review-capable
    bearer token. Unauthenticated: GET /health and GET /self-check.'
  version: 0.3.0
paths:
...
  /api/v1/investigations/{investigation_id}/cancel:
    post:
      tags:
      - Investigations
      summary: Cancel an investigation
      description: Request cancel. In-flight model generation may finish the current
        turn before stopping. Intended for automation (CI, scripts). Requires the
        **triage** capability.
      operationId: post_api_v1_investigations_investigation_id_cancel
      parameters:
      - name: investigation_id
        in: path
        required: true
        schema:
          type: string
          title: Investigation Id
      responses:
        '200':
          description: Cancel an investigation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvestigationStateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - shiftLeftToken: []
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InvestigationStateResponse:
      properties:
        investigation_id:
          type: string
          title: Investigation Id
        state:
          type: string
          title: State
        originating_investigation_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Originating Investigation Id
        resolved_commit_sha:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolved Commit Sha
      additionalProperties: true
      type: object
      required:
      - investigation_id
      - state
      title: InvestigationStateResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    shiftLeftToken:
      type: http
      scheme: bearer
      bearerFormat: slt_
      description: 'Operator API token minted by `./shift-left token mint` (example
        value `slt_EXAMPLE`, never a live token).


        Send `Authorization: Bearer slt_…` or `X-Shift-Left-Token: slt_…`. Actor identity
        always comes from the token; JSON bodies must not include `actor` or `approver`.


        Capabilities (also see docs/auth-tokens.md):

        - **review** — `POST /api/v1/review`

        - **triage** — finding status, non-BLOCK waivers, investigations

        - **approve** — grant/revoke approvals; BLOCK-level waivers

        - **admin** — all of the above, plus system settings and service lifecycle

        - **override** — block overrides when `policy.allow_block_override` is true

        - **deploy** — plan-only Terraform


        `admin` implies the others. The API listens on loopback of the operator host;
        a hosted OpenAPI page cannot call it.'
...
servers:
- url: http://127.0.0.1:8080
  description: Operator-host loopback. The orchestrator binds locally and is not reachable
    from GitHub Pages or other remote clients.